More function support.
This adds syntax for lambda x, y, z: expr type of expression. This can be used to create functions in GCL. Unfortunately to resolve ambiguity this removes support for ':' in tuple keys.
Also adds map() and filter() support, which can be used as a primitive form of loops.
Also adds support for the modulo (%) operator.
Also fixes a bug where a custom env wouldn't be propagated to includes.